// src/thunks/types.ts
export interface Extractor {
  name: string
  meta: MetaExtractor
}

export interface MetaExtractor {
  name: string
  taskflow: MetaTaskflow[]
  inputs: MetaItems[]
  preprocesses: MetaItems[]
  nodes: MetaNNNode[]
  postprocesses: MetaItems[]
  outputs: MetaItems[]
  extractors: MetaExtractor[]
}

export enum TaskflowType {
  UNSPECIFIED = 'unspecified',

  INPUT = 'input',

  PREPROCESS = 'preprocess',

  FORWARD = 'forward',

  BACKWARD = 'backward',

  POSTPROCESS = 'postprocess',

  OUTPUT = 'output',

  EXTRACTOR = 'extractor',
}
// ...
export interface MetaTaskflow {
  name: string
  the_type: TaskflowType
  flow_id: number

  value?: MetaItems | MetaNNNode | MetaExtractor

  op_type: OpType
  // biome-ignore lint/suspicious/noExplicitAny: op_params can be any
  op_params: any
}

export interface MetaItems {
  name: string
  items: MetaItem[]
}
// ...
export interface MetaNNNode {
  name: string
  inputs: MetaNNRecord[]
  params: MetaNNParameter[]
  activation?: MetaNNRecord | null
  gradient_inputs: MetaNNRecord[]
  gradient_params: MetaNNParameter[]
  gradients: MetaNNRecord[]
  children: MetaNNNode[] | number
}

export interface MetaNNParameter {
  name: string
  parameter: MetaNNRecord
}

export interface MetaNNRecord {
  name: string
  the_type: RecordType
  record: MetaNNTensor | MetaNNRecord[] | Record<string, MetaNNRecord>
  data_id: string
}

export type RecursiveNumber = number[] | RecursiveNumber[]

export enum NNTensorArrayType {
  FLOAT32 = 'float32',
  UINT8 = 'uint8',
}
// ...
export enum ProtobufType {
  Item = 'item',
  NNRecord = 'nnrecord',
  NNParameter = 'nnparameter',
  OpItem = 'op_item',
}

export interface ProtobufResponse {
  bytes: string
  errmsg: string
}

export interface ProtobufContextInfo {
  affineRAS: number[][]
  reducedDimsSAR: boolean[]
  dimsSAR: number[]
}

export interface ProtobufInfo {
  extractorID: string
  flowType: nnextractor_pb.TaskflowType
  flowID: number
  dataID: string

  dimsSAR: number[]

  contextInfo: ProtobufContextInfo

  protobufType: ProtobufType

  isGeoIdentity?: boolean

  isSegmentation?: boolean
}

export interface RefImgInfo {
  imgSizeSAR: number[]
  originRAS: number[]
  spacingRAS: number[]
  directionRAS: number[][]
  affineRAS: number[][]
  invAffineRAS: number[][]
}